Wildemount

+++ main/cogs/utils/db.py
@@ -45,11 +45,12 @@ async def create_schemas(conn):
     await conn.execute(sql)
     # ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
     # Custom tables for certain servers
-    sql = '''CREATE TABLE IF NOT EXISTS quest(server_id BIGINT NOT NULL,
-                                               quest_id BIGINT UNIQUE,
-                                               author TEXT NOT NULL,
-                                               quest_type text NOT NULL,
-                                               msg TEXT);'''
+    sql = '''CREATE TABLE IF NOT EXISTS quest(server_id INTEGER NOT NULL,
+                                              quest_id INTEGER UNIQUE,
+                                              author TEXT NOT NULL,
+                                              quest_type text NOT NULL,
+                                              msg TEXT);'''
+
 
     await conn.execute(sql)
     return
